+++ src/hex.test.ts
@@ -11,7 +11,7 @@ describe("encodeHex()", () => {
 	});
 });
 
-describe("Base32.decode()", () => {
+describe("Base32.decodeIgnorePadding()", () => {
 	test("Returns encoded data", () => {
 		const cases = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
 		for (const length of cases) {