+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # @oslojs/encoding
+
+## 0.3.0
+
+- [Breaking] Add `Encoding.encodeNoPadding()` and `Encoding.decodeIgnorePadding()`
+- [Breaking] Remove `includingPadding` option from `Base64Encoding.encode()` and `Base32Encoding.encode()`
+- [Breaking] Remove `strict` option from `Base64Encoding.decodeIgnorePadding()` and `Base32Encoding.decodeIgnorePadding()`
+- Add `Encoding.encodeNoPadding()`,  `Base64Encoding.encodeNoPadding()`,  `Base32Encoding.encodeNoPadding()`
+- Add `Encoding.decodeIgnorePadding()`,  `Base64Encoding.decodeIgnorePadding()`,  `Base32Encoding.decodeIgnorePadding()`